reorganizing folders and image files
