+++ frontend/components/styles/resourceCard.js
